Consolidate Utils to GameSimulatorUtils
