+++ src/features/gameSimulator/GameSimulatorUtils.ts
