--- src/pages/academy/gameSimulator/subcomponents/chapterPublisher/ChapterPublisherUtils.tsx
@@ -1,4 +1,0 @@
-export const dateOneYearFromNow = (date: Date) => {
-  date.setFullYear(date.getFullYear() + 1);
-  return date;
-};
